host: Move update related code into host/update.go
